+++ HISTORY.md
@@ -3,6 +3,7 @@
 
 ## not yet released, version 5.0.2
 
+- Fixed #1088: function arguments not being passed to `rawArgs` functions.
 - Fixed #1136: rocket trajectory example broken (since v4.0.0).
 - Fixed advanced example `add_new_datatypes`.
 - Upgraded to `typed-function@1.0.4`.